refactor: move go code into internal dir
