+++ internal/config/testing/alwaysPasses.rego
