+++ internal/config/testing/regoWithSyntaxError.rego
