+++ internal/rego/rego_test.go
