+++ internal/util/network.go
