+++ internal/util/network_test.go
